services/jaccard/requirements.txt
grpcio
protobuf
requests
fastapi
uvicorn
pydantic